refactor: LottoNumber에 대해 패키지 세분화
