+++ src/main/java/lotto/domain/lottoticket/lottonumber/LottoNumber.java
@@ -1,4 +1,4 @@
-package lotto.domain.lottoticket;
+package lotto.domain.lottoticket.lottonumber;
 
 import lotto.domain.exception.InvalidLottoNumberCreationException;
