+++ src/test/java/lotto/domain/lottoticket/LottoTicketsTest.java
@@ -1,5 +1,7 @@
 package lotto.domain.lottoticket;
 
+import lotto.domain.lottoticket.lottonumber.LottoNumber;
+import lotto.domain.lottoticket.lottonumber.LottoNumberPool;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
