+++ src/test/java/lotto/domain/lottoticket/lottonumber/LottoNumberPoolTest.java
@@ -1,4 +1,4 @@
-package lotto.domain.lottoticket;
+package lotto.domain.lottoticket.lottonumber;
 
 import org.junit.jupiter.api.Test;
